Add NaN check to compression
